LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-button-activation-content-hit-test-expected.txt



FAIL CSS Test: ::scroll-button() activation by clicking on text content area assert_equals: expected 604 but got 0
